Having base html and removing JQuery
