+++ log_analyzer/src/templates/index.html
@@ -1,68 +1,14 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
+{% extends 'base.html' %}
 
-    <script src="https://code.jquery.com/jquery-3.5.1.min.js"
-            integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/3.1.2/socket.io.js"
-            integrity="sha512-YybopSVjZU0fe8TY4YDuQbP5bhwpGBE/T6eBUEZ0usM72IWBfWrgVI13qfX4V2A/W7Hdqnm7PIOYOwP9YHnICw==
-            " crossorigin="anonymous"></script>
-    <!--  Bootstrap start  -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0"
-            crossorigin="anonymous"></script>
-    <!-- Bootstrap end  -->
+{% block js %}
+<script type="module" src="{{ url_for('static', filename='js/index.js') }}"></script>
+{% endblock %}
 
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.9.4/Chart.min.js" 
-            integrity="sha512-d9xgZrVZpmmQlfonhQUvTR7lMPtO7NkZMkA0ABN3PHCbKA5nqylQ/yWlFAyY6hYgdF1Qh6nYiuADWwKB4C2WSw==" 
-            crossorigin="anonymous"></script>
-
-    <script type="text/javascript" src="https://unpkg.com/vis-network/standalone/umd/vis-network.min.js"></script>
-
-    <script type="module" src="{{ url_for('static', filename='js/application.js') }}"></script>
-    <link rel="stylesheet" href="static/css/main.css"/>
-
-    <title>Log Analyzer</title>
-
-</head>
-<body>
-
-<nav class="navbar navbar-expand-md navbar-dark bg-dark fixed-top">
-  <div class="container-fluid">
-    <a class="navbar-brand" href="#">Log Analyzer</a>
-    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarsExampleDefault"
-            aria-controls="navbarsExampleDefault" aria-expanded="false" aria-label="Toggle navigation">
-    </button>
-
-    <div class="collapse navbar-collapse">
-      <ul class="navbar-nav navbar-center">
-        <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="/">Home</a>
-        </li>
-        <li class = "nav-item">
-            <a class="nav-link active" aria-current="page" href="#"> | </a>
-        </li>
-        <li class="nav-item">
-            <a class="nav-link" href="/stats">Stats</a>
-        </li>
-        <li class = "nav-item">
-          <a class="nav-link active" aria-current="page" href="#"> | </a>
-      </li>
-      <li class="nav-item">
-          <a class="nav-link" href="/settings">Settings</a>
-      </li>
-      </ul>
-    </div>
+{% block content %}
+  <div class="channels-container"></div>
+  <div class="container charts">
+    <canvas id="execution_time_chart"></canvas>
+    <canvas id="drop_rate_chart"></canvas>
   </div>
-</nav>
-</body>
-    <div class="channels-container"></div>
-    <div class="container charts">
-      <canvas id="execution_time_chart"></canvas>
-      <canvas id="drop_rate_chart"></canvas>
-    </div>
-    <div id="channels-map"></div>
-</html>
+  <div id="channels-map"></div>
+{% endblock %}
